Classifier_Modules/requirements.txt
gensim
mysql-connector
sklearn
tensorflow
keras
pika
networkx
pandas
tqdm
joblib